i18n proof-of-concept with react-intl
This is a proof of concept of i18n with the react-intl module.

Only part of the login form is translated. You can try this by going to http://localhost:8065/myteam?locale=de (assuming you have team called myteam and you are not currently logged in.)

The locale is determined by the URL in this POC, but it could be determined by request headers, cookies or user settings as well (or any mix of those.)